+++ src/main/java/org/github/turostaska/service/ICharacterService.java
@@ -26,7 +26,7 @@ public interface ICharacterService {
      * @param ID The desired player's ID
      * @return The player with the given ID if it exists, otherwise an empty object
      */
-    Optional<Player> getPlayerById(int ID);
+    Optional<Player> getPlayerById(Long ID);
 
     /**
      * Returns the player with the desired name if it exists, otherwise returns an empty object
@@ -67,7 +67,7 @@ public interface ICharacterService {
      * @param ID The desired NPC's ID
      * @return The NPC with the given ID if it exists, otherwise an empty object
      */
-    Optional<NPC> getNPCById(int ID);
+    Optional<NPC> getNPCById(Long ID);
 
     /**
      * Returns the NPC with the desired name if it exists, otherwise returns an empty object
