+++ src/main/java/org/github/turostaska/service/IUserService.java
@@ -24,7 +24,7 @@ public interface IUserService {
      * @param ID The desired user's ID
      * @return The user with the given ID if it exists, otherwise an empty object
      */
-    Optional<User> getById(int ID);
+    Optional<User> getById(Long ID);
 
     /**
      * Returns the user with the desired username if it exists, otherwise returns an empty object
