+++ src/main/java/org/github/turostaska/service/impl/collection/CollectionTechniqueService.java
@@ -27,7 +27,7 @@ public class CollectionTechniqueService implements ITechniqueService {
     }
 
     @Override
-    public Optional<Technique> getById(int ID) {
+    public Optional<Technique> getById(Long ID) {
         return dao.getById(ID);
     }
